requirements.txt
streamlit
pandas
openpyxl==3.1.2
python-docx==0.8.11
fpdf2==2.7.4
plotly
google-analytics-data==0.17.0
requests
